requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
google-auth==2.39.0
google-genai==1.13.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.57
langchain-openai==0.3.15
langchain-text-splitters==0.3.8
langfuse==2.60.3
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-checkpoint-mongodb==0.1.3
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.40
motor==3.7.0
openai==1.76.2
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.11.3
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
zstandard==0.23.0
